pkg/util: move proxy into k8s.io/apiserver
