+++ staging/src/k8s.io/apiserver/pkg/util/proxy/doc.go
@@ -15,4 +15,4 @@ limitations under the License.
 */
 
 // Package proxy provides transport and upgrade support for proxies
-package proxy // import "k8s.io/kubernetes/pkg/util/proxy"
+package proxy // import "k8s.io/apiserver/pkg/util/proxy"
